run-kgr-test.sh: Check if test_script.sh exists before running the test
Otherwise kgr-test returns saying that nothing is being done.

+++ scripts/run-kgr-test.sh
@@ -10,6 +10,12 @@ fi
 # directory of the current script is the bsc number
 bsc=$(basename "$(pwd)")
 
+TEST_SCRIPT="repro/${bsc}_test_script.sh"
+if [ ! -f "$TEST_SCRIPT" ]; then
+	echo "Missing $TEST_SCRIPT. Aborting."
+	exit 1
+fi
+
 # Remove any previous test output
 rm -rf tests.out/*
